% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_governorates}
\alias{plot_governorates}
\title{Plot a map of Tunisia showing the boundaries of Governorates.}
\usage{
plot_governorates()
}
\value{
ggplot2 plot
}
\description{
Plot a map of Tunisia showing the boundaries of Governorates.
}
