% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_municipalities}
\alias{plot_municipalities}
\title{Plot a map of Tunisia showing the boundaries of Municipalities.}
\usage{
plot_municipalities()
}
\value{
ggplot2 plot
}
\description{
Plot a map of Tunisia showing the boundaries of Municipalities.
}
