% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tun_governorates}
\alias{tun_governorates}
\title{Map of Tunisia divided by Governorates.}
\format{
A data frame with `r nrow(tunmaps::tun_governorates)` rows
\describe{
  \item{geometry}{Polygons representing Governorates}
  \item{id_gov}{Administrative Identifier of Governorate}
  \item{name_gov_fr}{Name of Governorate in French}
}
}
\usage{
tun_governorates
}
\description{
The Map of Tunisia divided by Governorates with
their administrative identifiers and their French names.
}
\keyword{datasets}
