% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tun_municipalities}
\alias{tun_municipalities}
\title{Map of Tunisia divided by Municipalities.}
\format{
A data frame with `r nrow(tunmaps::tun_municipalities)` rows
\describe{
  \item{geometry}{Polygons representing Municipalities}
  \item{id_gov}{Administrative Identifier of Governorate}
  \item{name_gov_fr}{Name of Governorate in French}
}
}
\usage{
tun_municipalities
}
\description{
The Map of Tunisia divided by Municipalities with
their administrative identifiers and their French names.
}
\keyword{datasets}
